Update README.md
This project is a web application built with React.js and TypeScript. It allows you to view, edit, and manage details of some of the most famous celebrities in the world. Features :- A search bar to quickly find celebrities by name. Clickable celebrity profiles (accordion-style), where only one profile expands at a time. Options to edit or delete details, but only if the celebrity is an adult. Validation to make sure you can not enter invalid data like numbers in a name or blank fields. Editable fields like gender (dropdown), country (text), and description (textarea).